requirements.txt
boto>=2.14.0
PyYAML>=3.10
moto==0.2.9
mock==1.0.1
